+++ zcash_proofs/CHANGELOG.md
@@ -12,7 +12,7 @@ workspace.
 
 ### Changed
 - MSRV is now 1.85.1.
-- Migrated to `sapling-crypto 0.6`.
+- Migrated to `sapling-crypto 0.7`, `zcash_primitives 0.27`.
 
 ## [0.26.1] - 2025-10-18